+++ src/main/java/com/cellterminal/gui/tab/StorageBusPartitionTabController.java
@@ -64,6 +64,7 @@ public class StorageBusPartitionTabController implements ITabController {
         lines.add(I18n.format("gui.cellterminal.controls.jei_drag"));
         lines.add(I18n.format("gui.cellterminal.controls.click_to_remove"));
         lines.add(I18n.format("gui.cellterminal.controls.double_click_storage"));
+        lines.add(I18n.format("gui.cellterminal.right_click_rename"));
 
         return lines;
     }