+++ src/main/java/com/cellterminal/proxy/ClientProxy.java
@@ -9,6 +9,7 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import com.cellterminal.ItemRegistry;
 import com.cellterminal.client.BlockHighlightRenderer;
 import com.cellterminal.client.KeyBindings;
+import com.cellterminal.client.KeyInputHandler;
 import com.cellterminal.client.UpgradeTooltipHandler;
 import com.cellterminal.integration.AE2WUTIntegration;
 
@@ -40,6 +41,9 @@ public class ClientProxy extends CommonProxy {
 
         // Register upgrade tooltip handler
         MinecraftForge.EVENT_BUS.register(new UpgradeTooltipHandler());
+
+        // Register key input handler for world-context keybinds (wireless terminal)
+        MinecraftForge.EVENT_BUS.register(new KeyInputHandler());
 
         // Register with AE2WUT client-side if present
         AE2WUTIntegration.registerGui();