+++ CHANGELOG.md
@@ -8,9 +8,19 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
 
+## [1.4.0-rc2] - 2026-02-19
+### Added
+- Add right-click rename for drives/ME chests, storage cells, and storage buses on all tabs (1-5).
+- Add keybind (default: Shift+D) to open the Wireless Cell Terminal from anywhere. Compatible with both AE2's Wireless Terminal and AE2 Wireless Universal Terminal.
+
+### Fixed
+- Fix the terminal not properly refreshing when opening it on a subnet.
+
+
 ## [1.4.0-rc1] - 2026-02-18
 ### Added
 - Add subnet overview screen, showing all connected subnets and allowing to load them in the main terminal view. This allows better management of multiple subnets, instead of having a different terminal for each subnet and no way to see the overall structure of the network.
+- Add guiding arrows to all highlights above 50 blocks away, for as long as the highlight is active. The arrow size and arrow text size can be be configured in the client config.
 
 
 ## [1.3.0] - 2026-02-16